feat: 添加V100 FP16训练支持及稳定性优化
refactor(训练脚本): 重构DeepSpeed配置处理逻辑
fix(模型): 修复FP16训练下的数值稳定性问题
docs: 添加安装指南和DeepSpeed训练文档
test: 增加溢出检测和梯度调试功能

+++ src/openpi/models_pytorch/pi05_subtask.py
@@ -153,8 +153,8 @@ class PI05SubtaskPytorch(PI0Pytorch):
             v_t = result["v_t"]
             ce_loss = result["ce_loss"].mean()
 
-            # Flow matching loss
-            flow_loss = F.mse_loss(u_t, v_t, reduction="mean")
+            # Compute flow loss in fp32 for numerical stability under fp16 training.
+            flow_loss = F.mse_loss(u_t.float(), v_t.float(), reduction="mean")
 
             # Combined loss: CE + alpha * flow_matching (Equation 1)
             combined_loss = ce_loss + self.alpha * flow_loss
@@ -176,7 +176,7 @@ class PI05SubtaskPytorch(PI0Pytorch):
                 x_t=x_t,
                 time=time,
             )
-            flow_loss = F.mse_loss(u_t, v_t, reduction="mean")
+            flow_loss = F.mse_loss(u_t.float(), v_t.float(), reduction="mean")
 
             return {
                 "loss": flow_loss,
